requirements.txt
###### Requirements without Version Specifiers ######
camelot-py
opencv-python
ghostscript
pandas
pyyaml